requirements.txt
anyio==3.6.1
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
passlib==1.7.4
psycopg2-binary==2.9.3
pydantic==1.9.1
PyJWT==2.4.0
python-decouple==3.6
python-dotenv==0.20.0
python-multipart==0.0.5
requests==2.28.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing-extensions==4.2.0
urllib3==1.26.10
uvicorn==0.17.6
